query 4 concluida(?)

+++ trabalho-pratico/Include/statistics.h
@@ -22,6 +22,6 @@ double city_get_money(void *stats_d,char *ID);
 
 double city_get_money_notip(void *stats_p,char *ID);
 
-short city_get_num_rides(void *stats_d,char *ID);
+int city_get_num_rides(void *stats_d,char *ID);
 
 #endif
